cleanup: extract transaction schema and form fields into separate modules
Split add-transaction.tsx (~400 lines) into three focused modules:
- transaction-schema.ts: Zod validation schemas, types, and config
- transaction-form-fields.tsx: shared form field components
- add-transaction.tsx: dialog wrapper with form state and submission
