+++ src/components/forms/add-transaction.tsx
@@ -3,15 +3,10 @@
 import { useState, useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import * as z from 'zod';
-import { format } from 'date-fns';
-import { CalendarIcon, Plus } from 'lucide-react';
+import { Plus } from 'lucide-react';
 import { Decimal } from 'decimal.js';
 
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
 import {
   Dialog,
   DialogContent,
@@ -21,21 +16,7 @@ import {
   DialogTrigger,
   DialogFooter,
 } from '@/components/ui/dialog';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover';
-import { Calendar } from '@/components/ui/calendar';
 import { Badge } from '@/components/ui/badge';
-import { cn } from '@/lib/utils';
 import { useTransactionStore, usePortfolioStore } from '@/lib/stores';
 import {
   showSuccessNotification,
@@ -47,250 +28,12 @@ import {
   RSUTransactionMetadata,
 } from '@/types';
 import { assetQueries } from '@/lib/db';
-import { ESPPTransactionFormFields } from './espp-transaction-form';
-import { RSUTransactionFormFields } from './rsu-transaction-form';
-import { FormProvider } from 'react-hook-form';
-
-const baseTransactionSchema = z.object({
-  type: z.enum([
-    'buy',
-    'sell',
-    'dividend',
-    'interest',
-    'split',
-    'transfer_in',
-    'transfer_out',
-    'fee',
-    'tax',
-    'spinoff',
-    'merger',
-    'reinvestment',
-    'espp_purchase',
-    'rsu_vest',
-    'deposit',
-    'withdrawal',
-    'liability_payment',
-  ]),
-  assetSymbol: z
-    .string()
-    .min(1, 'Asset symbol is required')
-    .max(10, 'Symbol too long'),
-  assetName: z.string().optional(),
-  date: z.date({
-    required_error: 'Transaction date is required',
-  }),
-  quantity: z.string().refine(
-    (val) => {
-      const num = parseFloat(val);
-      return !isNaN(num) && num > 0;
-    },
-    {
-      message: 'Quantity must be a positive number',
-    }
-  ),
-  price: z.string().refine(
-    (val) => {
-      const num = parseFloat(val);
-      return !isNaN(num) && num >= 0;
-    },
-    {
-      message: 'Price must be a non-negative number',
-    }
-  ),
-  fees: z
-    .string()
-    .optional()
-    .refine(
-      (val) => {
-        if (!val || val === '') return true;
-        const num = parseFloat(val);
-        return !isNaN(num) && num >= 0;
-      },
-      {
-        message: 'Fees must be a non-negative number',
-      }
-    ),
-  notes: z.string().max(500, 'Notes too long').optional(),
-  // ESPP-specific fields (conditional)
-  grantDate: z.date().optional(),
-  marketPriceAtGrant: z.string().optional(),
-  marketPriceAtPurchase: z.string().optional(),
-  discountPercent: z.string().optional(),
-  // RSU-specific fields (conditional)
-  vestingDate: z.date().optional(),
-  grossSharesVested: z.string().optional(),
-  sharesWithheld: z.string().optional(),
-  vestingPrice: z.string().optional(),
-  taxWithheldAmount: z.string().optional(),
-});
-
-// Refined schema with conditional validation
-const transactionSchema = baseTransactionSchema
-  .refine(
-    (data) => {
-      if (data.type !== 'espp_purchase') return true;
-      return data.grantDate !== undefined;
-    },
-    {
-      message: 'Grant date is required for ESPP transactions',
-      path: ['grantDate'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'espp_purchase') return true;
-      if (!data.marketPriceAtGrant) return false;
-      const num = parseFloat(data.marketPriceAtGrant);
-      return !isNaN(num) && num > 0;
-    },
-    {
-      message: 'Market price at grant is required for ESPP transactions',
-      path: ['marketPriceAtGrant'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'espp_purchase') return true;
-      if (!data.marketPriceAtPurchase) return false;
-      const num = parseFloat(data.marketPriceAtPurchase);
-      return !isNaN(num) && num > 0;
-    },
-    {
-      message: 'Market price at purchase is required for ESPP transactions',
-      path: ['marketPriceAtPurchase'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'espp_purchase') return true;
-      if (!data.discountPercent) return false;
-      const num = parseFloat(data.discountPercent);
-      return !isNaN(num) && num >= 0 && num <= 100;
-    },
-    {
-      message: 'Discount percentage must be between 0 and 100',
-      path: ['discountPercent'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'espp_purchase') return true;
-      if (!data.grantDate || !data.date) return true;
-      return data.grantDate < data.date;
-    },
-    {
-      message: 'Grant date must be before purchase date',
-      path: ['grantDate'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'rsu_vest') return true;
-      return data.vestingDate !== undefined;
-    },
-    {
-      message: 'Vesting date is required for RSU transactions',
-      path: ['vestingDate'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'rsu_vest') return true;
-      if (!data.grossSharesVested) return false;
-      const num = parseFloat(data.grossSharesVested);
-      return !isNaN(num) && num > 0;
-    },
-    {
-      message: 'Gross shares vested is required for RSU transactions',
-      path: ['grossSharesVested'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'rsu_vest') return true;
-      if (!data.sharesWithheld) return false;
-      const num = parseFloat(data.sharesWithheld);
-      return !isNaN(num) && num >= 0;
-    },
-    {
-      message: 'Shares withheld must be 0 or greater',
-      path: ['sharesWithheld'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'rsu_vest') return true;
-      if (!data.vestingPrice) return false;
-      const num = parseFloat(data.vestingPrice);
-      return !isNaN(num) && num > 0;
-    },
-    {
-      message: 'Vesting price is required for RSU transactions',
-      path: ['vestingPrice'],
-    }
-  )
-  .refine(
-    (data) => {
-      if (data.type !== 'rsu_vest') return true;
-      if (!data.grossSharesVested || !data.sharesWithheld) return true;
-      const gross = parseFloat(data.grossSharesVested);
-      const withheld = parseFloat(data.sharesWithheld);
-      return withheld <= gross;
-    },
-    {
-      message: 'Shares withheld cannot exceed gross shares vested',
-      path: ['sharesWithheld'],
-    }
-  );
-
-type TransactionFormValues = z.infer<typeof transactionSchema>;
-
-const transactionTypes = [
-  {
-    value: 'buy',
-    label: 'Buy',
-    color: 'bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-300',
-  },
-  {
-    value: 'sell',
-    label: 'Sell',
-    color: 'bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-300',
-  },
-  {
-    value: 'dividend',
-    label: 'Dividend',
-    color: 'bg-blue-100 text-blue-800 dark:bg-blue-900 dark:text-blue-300',
-  },
-  {
-    value: 'split',
-    label: 'Stock Split',
-    color:
-      'bg-purple-100 text-purple-800 dark:bg-purple-900 dark:text-purple-300',
-  },
-  {
-    value: 'espp_purchase',
-    label: 'ESPP Purchase',
-    color: 'bg-cyan-100 text-cyan-800 dark:bg-cyan-900 dark:text-cyan-300',
-  },
-  {
-    value: 'rsu_vest',
-    label: 'RSU Vest',
-    color:
-      'bg-indigo-100 text-indigo-800 dark:bg-indigo-900 dark:text-indigo-300',
-  },
-  {
-    value: 'transfer_in',
-    label: 'Transfer In',
-    color:
-      'bg-yellow-100 text-yellow-800 dark:bg-yellow-900 dark:text-yellow-300',
-  },
-  {
-    value: 'transfer_out',
-    label: 'Transfer Out',
-    color:
-      'bg-orange-100 text-orange-800 dark:bg-orange-900 dark:text-orange-300',
-  },
-];
+import {
+  transactionSchema,
+  transactionTypes,
+  type TransactionFormValues,
+} from './transaction-schema';
+import { TransactionFormFields } from './transaction-form-fields';
 
 interface TransactionDialogProps {
   mode?: 'create' | 'edit';
@@ -354,12 +97,10 @@ function TransactionDialog({
             price: transaction.price.toString(),
             fees: transaction.fees.toString(),
             notes: transaction.notes || '',
-            // ESPP fields
             grantDate: undefined,
             marketPriceAtGrant: '',
             marketPriceAtPurchase: '',
             discountPercent: '',
-            // RSU fields
             vestingDate: undefined,
             grossSharesVested: '',
             sharesWithheld: '',
@@ -375,12 +116,10 @@ function TransactionDialog({
             price: '',
             fees: '0',
             notes: '',
-            // ESPP fields
             grantDate: undefined,
             marketPriceAtGrant: '',
             marketPriceAtPurchase: '',
             discountPercent: '',
-            // RSU fields
             vestingDate: undefined,
             grossSharesVested: '',
             sharesWithheld: '',
@@ -391,13 +130,11 @@ function TransactionDialog({
   });
 
   const {
-    register,
     handleSubmit,
-    formState: { errors, isValid },
+    formState: { isValid },
     reset,
     setValue,
     watch,
-    trigger: triggerValidation,
   } = form;
 
   // Load asset symbol when dialog opens in edit mode
@@ -436,8 +173,6 @@ function TransactionDialog({
   }, [open, mode, transaction, setValue]);
 
   const watchedType = watch('type');
-  const watchedDate = watch('date');
-  const watchedSymbol = watch('assetSymbol');
   const watchedQuantity = watch('quantity');
   const watchedPrice = watch('price');
   const watchedFees = watch('fees');
@@ -475,7 +210,7 @@ function TransactionDialog({
       // Handle race condition - asset might have been created between check and create
       const retryExisting = await assetQueries.getBySymbol(symbol);
       if (retryExisting) return retryExisting;
-      throw error; // Re-throw if it's a different error
+      throw error;
     }
   };
 
@@ -490,7 +225,6 @@ function TransactionDialog({
 
     setIsSubmitting(true);
     try {
-      // Build common transaction fields
       const quantity = new Decimal(data.quantity);
       const price = new Decimal(data.price);
       const transactionData: Omit<Transaction, 'id' | 'portfolioId' | 'assetId' | 'currency'> = {
@@ -621,208 +355,10 @@ function TransactionDialog({
             </DialogHeader>
 
             <form onSubmit={handleSubmit(onSubmit)} className="space-y-6">
-              {/* Transaction Type */}
-              <div className="space-y-2">
-                <Label htmlFor="type">Transaction Type *</Label>
-                <Select
-                  value={watchedType}
-                  onValueChange={(value: string) =>
-                    setValue('type', value as TransactionFormValues['type'], { shouldValidate: true })
-                  }
-                >
-                  <SelectTrigger>
-                    <SelectValue placeholder="Select transaction type" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {transactionTypes.map((type) => (
-                      <SelectItem key={type.value} value={type.value}>
-                        <div className="flex items-center gap-2">
-                          <Badge className={type.color} variant="secondary">
-                            {type.label}
-                          </Badge>
-                        </div>
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-                {errors.type && (
-                  <p className="text-sm text-red-600">{errors.type.message}</p>
-                )}
-              </div>
-
-              {/* Asset Information */}
-              <div className="grid grid-cols-2 gap-4">
-                <div className="space-y-2">
-                  <Label htmlFor="assetSymbol">Asset Symbol *</Label>
-                  <Input
-                    id="assetSymbol"
-                    placeholder="AAPL, BTC, SPY"
-                    {...register('assetSymbol')}
-                    onChange={(e) => {
-                      const value = e.target.value.toUpperCase();
-                      setValue('assetSymbol', value, { shouldValidate: true });
-                    }}
-                    className={errors.assetSymbol ? 'border-red-500' : ''}
-                  />
-                  {errors.assetSymbol && (
-                    <p className="text-sm text-red-600">
-                      {errors.assetSymbol.message}
-                    </p>
-                  )}
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="assetName">Asset Name (Optional)</Label>
-                  <Input
-                    id="assetName"
-                    placeholder="Apple Inc."
-                    {...register('assetName')}
-                  />
-                </div>
-              </div>
-
-              {/* Date */}
-              <div className="space-y-2">
-                <Label>Transaction Date *</Label>
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <Button
-                      variant="outline"
-                      className={cn(
-                        'w-full justify-start text-left font-normal',
-                        !watchedDate && 'text-muted-foreground',
-                        errors.date && 'border-red-500'
-                      )}
-                    >
-                      <CalendarIcon className="mr-2 h-4 w-4" />
-                      {watchedDate ? format(watchedDate, 'PPP') : 'Pick a date'}
-                    </Button>
-                  </PopoverTrigger>
-                  <PopoverContent className="w-auto p-0" align="start">
-                    <div className="p-4">
-                      <Input
-                        type="date"
-                        value={
-                          watchedDate ? format(watchedDate, 'yyyy-MM-dd') : ''
-                        }
-                        onChange={(e) => {
-                          if (e.target.value) {
-                            setValue('date', new Date(e.target.value), {
-                              shouldValidate: true,
-                            });
-                          }
-                        }}
-                        max={format(new Date(), 'yyyy-MM-dd')}
-                        min="1900-01-01"
-                      />
-                    </div>
-                  </PopoverContent>
-                </Popover>
-                {errors.date && (
-                  <p className="text-sm text-red-600">{errors.date.message}</p>
-                )}
-              </div>
-
-              {/* Quantity and Price */}
-              <div className="grid grid-cols-2 gap-4">
-                <div className="space-y-2">
-                  <Label htmlFor="quantity">
-                    Quantity *{watchedType === 'dividend' && ' (Shares held)'}
-                  </Label>
-                  <Input
-                    id="quantity"
-                    type="number"
-                    step="any"
-                    placeholder="100"
-                    {...register('quantity')}
-                    className={errors.quantity ? 'border-red-500' : ''}
-                  />
-                  {errors.quantity && (
-                    <p className="text-sm text-red-600">
-                      {errors.quantity.message}
-                    </p>
-                  )}
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="price">
-                    {watchedType === 'dividend'
-                      ? 'Dividend per Share *'
-                      : 'Price per Share *'}
-                  </Label>
-                  <Input
-                    id="price"
-                    type="number"
-                    step="0.01"
-                    placeholder="150.00"
-                    {...register('price')}
-                    className={errors.price ? 'border-red-500' : ''}
-                  />
-                  {errors.price && (
-                    <p className="text-sm text-red-600">
-                      {errors.price.message}
-                    </p>
-                  )}
-                </div>
-              </div>
-
-              {/* Fees and Total */}
-              <div className="grid grid-cols-2 gap-4">
-                <div className="space-y-2">
-                  <Label htmlFor="fees">Fees & Commissions</Label>
-                  <Input
-                    id="fees"
-                    type="number"
-                    step="0.01"
-                    placeholder="0.00"
-                    {...register('fees')}
-                    className={errors.fees ? 'border-red-500' : ''}
-                  />
-                  {errors.fees && (
-                    <p className="text-sm text-red-600">
-                      {errors.fees.message}
-                    </p>
-                  )}
-                </div>
-
-                <div className="space-y-2">
-                  <Label>Total Transaction Value</Label>
-                  <div className="rounded-md bg-muted p-3">
-                    <span className="text-lg font-semibold">
-                      ${calculateTotal().toFixed(2)}
-                    </span>
-                  </div>
-                </div>
-              </div>
-
-              {/* ESPP-specific fields */}
-              {watchedType === 'espp_purchase' && (
-                <FormProvider {...form}>
-                  <ESPPTransactionFormFields />
-                </FormProvider>
-              )}
-
-              {/* RSU-specific fields */}
-              {watchedType === 'rsu_vest' && (
-                <FormProvider {...form}>
-                  <RSUTransactionFormFields />
-                </FormProvider>
-              )}
-
-              {/* Notes */}
-              <div className="space-y-2">
-                <Label htmlFor="notes">Notes (Optional)</Label>
-                <Textarea
-                  id="notes"
-                  placeholder="Additional notes about this transaction..."
-                  rows={3}
-                  {...register('notes')}
-                  className={errors.notes ? 'border-red-500' : ''}
-                />
-                {errors.notes && (
-                  <p className="text-sm text-red-600">{errors.notes.message}</p>
-                )}
-              </div>
+              <TransactionFormFields
+                form={form}
+                calculateTotal={calculateTotal}
+              />
 
               <DialogFooter className="gap-2">
                 <Button
